app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["ng3rdstmadgke <ng3rdstmadgke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = "^0.110.2"
jinja2 = "^3.1.3"
google-api-python-client = "^2.126.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
